#28 内部DBのwebコンソールアクセスを提供する
H2DBのコンソールURLのセキュリティ設定とその他のパスのセキュリティ設定が共存できていなかったので修正しました。(H2DBのコンソールURLのセキュリティ設定のみ有効になっていた)
