--- webapps/homepage/src/main/kotlin/jp/kodaira/boardgame/homepage/properties/H2ConsoleProperties.kt
@@ -1,8 +1,0 @@
-package jp.kodaira.boardgame.homepage.properties
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.stereotype.Component
-
-@Component
-@ConfigurationProperties(prefix = "spring.h2.console")
-data class H2ConsoleProperties(var path: String = "")
